% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alpha_Testing_Function.R
\name{AT_mr_sf_mp}
\alias{AT_mr_sf_mp}
\title{Alpha Testing -
Multipe Return Horizon, Single Factor, Multiple Periods}
\usage{
AT_mr_sf_mp(
  data = NULL,
  return_cols = NULL,
  fname = NULL,
  fftile = 5,
  standardize = T
)
}
\arguments{
\item{data}{dataframe containing return column and factor value column}

\item{return_cols}{vector of column names of returns}

\item{fname}{character column name of factor}

\item{fftile}{integer number of fractiles to use in spliting data}
}
\description{
Runs AT_sr_sf_mp for multiple return horizons
}
\details{
UPDATED - This function returns for the universe and each quintile
 for each period and for each return horizon: \cr
      number of observations\cr
      average return\cr
      average relative return for quintiles\cr
      security level hit rate of universe vs 0\cr
      security level hit rate of universe vs universe average\cr
      information coefficient based on Z-Score of factor\cr
      information coefficient based on rank of factor\cr
      quintile spread between top and bottom quintile\cr
}
