% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alpha_Testing_Function.R
\name{AT_sr_sf_mp}
\alias{AT_sr_sf_mp}
\title{Alpha Testing -
Single Return Horizon, Single Factor, Multiple Periods}
\usage{
AT_sr_sf_mp(data = NULL, fname = NULL, rname = NULL, fftile = 5)
}
\arguments{
\item{data}{dataframe containing return column and factor value column}

\item{fname}{character column name of factor}

\item{rname}{character column name of return}

\item{fftile}{integer number of fractiles to use in spliting data}
}
\description{
Runs AT_sr_sf_sp for multiple periods
}
\details{
This function returns for the universe and each quintile for each period: \cr
      number of observations\cr
      average return\cr
      average relative return for quantiles\cr
      security level hit rate of universe vs 0\cr
      security level hit rate of universe vs universe average\cr
      information coefficient based on Z-Score of factor\cr
      information coefficient based on rank of factor\cr
      quintile spread between top and bottom quintile\cr
}
