% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alpha_Testing_Function.R
\name{AT_sr_sf_sp}
\alias{AT_sr_sf_sp}
\title{Alpha Testing - Single Return Horizon, Single Factor, Single Period}
\usage{
AT_sr_sf_sp(data, fname, rname, fftile)
}
\arguments{
\item{data}{dataframe containing return column and factor value column}

\item{fname}{character column name of factor}

\item{rname}{character column name of return}

\item{fftile}{integer number of fractiles to use in spliting data}
}
\value{
None
}
\description{
This is the most baseline AT function.
}
\details{
This function returns for the universe and each quintile:
     number of observations \cr
     average return \cr
     average relative return for quintiles \cr
     security level hit rate of universe vs 0 \cr
     security level hit rate of universe vs universe average \cr
     information coefficient based on Z-Score of factor \cr
     information coefficient based on rank of factor \cr
     quintile spread between top and bottom quintile
}
\examples{
AT_sr_sf_sp()
<<<<<<< HEAD
=======

>>>>>>> 1fe3a938b2fd751077cf0e5d9cae153c87343613
}
