% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bloomberg_Custom_Index_Functions.R
\name{bb_customindex_aggregate_value}
\alias{bb_customindex_aggregate_value}
\title{Bloomberg Custom Index - Aggregate Value}
\usage{
bb_customindex_aggregate_value(Portfolio, Field, Overrides)
}
\arguments{
\item{Portfolio}{Dataframe containing Date, Bloomberg Security Identifier, Share Position}

\item{Field}{Character Bloomberg Field}

\item{Overrides}{Named Character Vector containing RBlpapi Overrides for Field}
}
\value{

}
\description{
Function to calcualted aggregate value for a portfolio
}
