% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bloomberg_Custom_Index_Functions.R
\name{bb_customindex_calculate_divisor}
\alias{bb_customindex_calculate_divisor}
\title{Bloomberg Custom Index - Index Divisor}
\usage{
bb_customindex_calculate_divisor(Portfolio)
}
\arguments{
\item{Portfolio}{Dataframe containing Date, Bloomberg Security Identifier, Share Position}
}
\value{

}
\description{
Function to calcualted index divisor for list of securities
}
\details{
Will enter
}
