% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bloomberg_Portfolio_Holdings.R
\name{bb_download_etf_holdings}
\alias{bb_download_etf_holdings}
\title{Bloomberg ETF Holdings Download using DDE}
\usage{
bb_download_etf_holdings(ETF = "ARKK US Equity")
}
\arguments{
\item{ETF}{character column name of ETF}
}
\value{
data
}
\description{
Download ETF HLDR Page usign DDE
}
