% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bloomberg_Portfolio_Holdings.R
\name{bb_download_portfolio_equity_deltaexp}
\alias{bb_download_portfolio_equity_deltaexp}
\title{Bloomberg Portfolio Holdings Download}
\usage{
bb_download_portfolio_equity_deltaexp(
  date = Sys.Date(),
  Portfolio = "U22936327-2 Client"
)
}
\arguments{
\item{date}{Date in Date format}

\item{Portfolio}{character column name of Portfolio}
}
\value{
data
}
\description{
Download Client Portfolio Holdings for Date
}
