% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bloomberg_Custom_Index_Functions.R
\name{bbcustomindex_aggregate_value_fa}
\alias{bbcustomindex_aggregate_value_fa}
\title{Bloomberg Custom Index - Aggregate Value (Data from Filings)}
\usage{
bbcustomindex_aggregate_value_fa(Portfolio, Field, Overrides)
}
\arguments{
\item{Portfolio}{Dataframe containing Date, Bloomberg Security Identifier, Share Position}

\item{Field}{Character Bloomberg Field}

\item{Overrides}{Named Character Vector containing RBlpapi Overrides for Field}
}
\value{

}
\description{
Function to calcualted aggregate value for a portfolio using filing data
}
