% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bloomberg_Custom_Index_Functions.R
\name{bbcustomindex_aggregate_valuepershare}
\alias{bbcustomindex_aggregate_valuepershare}
\title{Bloomberg Custom Index - Value Per Share}
\usage{
bbcustomindex_aggregate_valuepershare(Portfolio, Field, Overrides)
}
\arguments{
\item{Portfolio}{Dataframe containing Date, Bloomberg Security Identifier, Share Position}

\item{Field}{Character Bloomberg Field}

\item{Overrides}{Named Character Vector containing RBlpapi Overrides for Field}
}
\value{

}
\description{
Function to calcualted value per share for list of securities
}
