% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugReportServer.R
\name{bugReportServer}
\alias{bugReportServer}
\title{Bug Report Module Server}
\usage{
bugReportServer(
  id,
  data_files,
  session,
  app_name = "App",
  email = "thierry.worch@frieslandcampina.com",
  exclude_inputs = NULL,
  inputs_format = c("excel", "txt")
)
}
\arguments{
\item{id}{Module namespace ID}

\item{data_files}{Reactive expression returning file input(s)}

\item{session}{Parent module's session object}

\item{app_name}{Name for the output files and metadata}

\item{email}{Email address for bug reports}

\item{exclude_inputs}{Character vector of input names to exclude from capture}

\item{inputs_format}{Format for inputs file: "excel" or "txt"}
}
\value{
Something
}
\description{
Bug Report Module Server
}
\examples{
NULL
}
