% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugReportUI.R
\name{bugReportUI}
\alias{bugReportUI}
\title{Bug Report Module UI}
\usage{
bugReportUI(
  id,
  email = "thierry.worch@frieslandcampina.com",
  app_name = "App",
  button_label = "Download Bug Report",
  button_class = "xls_dl",
  button_width = "100\%"
)
}
\arguments{
\item{id}{Module namespace ID}

\item{email}{Email address for bug reports}

\item{app_name}{Name of the app/module for display}

\item{button_label}{Label for the download button}

\item{button_class}{CSS class for the download button}

\item{button_width}{Width of the download button}
}
\value{
Something
}
\description{
Bug Report Module UI
}
\examples{
NULL
}
