% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccmf_bloomberg_portfolio}
\alias{ccmf_bloomberg_portfolio}
\title{CCMF Portfolio Data for Bloomberg Custom Index Examples}
\format{
A data frame with 934 rows and 3 columns
\describe{
    \item{Date}{Date, in date format}
    \item{Security}{Bloomberg Ticker}
    \item{Position}{Position size in security}
}
}
\usage{
data(ccmf_bloomberg_portfolio)
}
\description{
CCMF Portfolio Data for Bloomberg Custom Index Examples
}
\keyword{datasets}
