% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ_import.R
\name{EQ_import}
\alias{EQ_import}
\title{Imports and Decode EyeQuestion Format Excel Sheets}
\usage{
EQ_import(
  file_path,
  product = TRUE,
  prod_order = NULL,
  attribute = TRUE,
  assessor = FALSE,
  convert2num = FALSE
)
}
\arguments{
\item{file_path}{The path to the file to import}

\item{product}{Boolean: should the Products be decoded?}

\item{prod_order}{NULL by default: provide the products' order in the factor}

\item{attribute}{Boolean: should the Attributes be decoded?}

\item{assessor}{Boolean: should the Assessors be decoded?}

\item{convert2num}{Boolean: converts variables to numerical (when Min/Max is not empty in the Attributes tab)}
}
\value{
The re-coded dataset
}
\description{
This function imports and decodes EQ files to be directly used in R or in Shiny Apps.
}
\examples{
NULL
}
