% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alpha_Testing_Function.R
\name{f_scoring}
\alias{f_scoring}
\title{Quantile & Z-Scoring}
\usage{
f_scoring(data, fname, fftile)
}
\arguments{
\item{data}{dataframe containing column with data to be scored}

\item{fname}{character column name of factor}

\item{fftile}{integer number of fractiles to use in spliting data}
}
\value{
data
}
\description{
Add Windsorized Z-Score and Quantile Score to Data
}
\details{
Used in Alpha Testing Functions
}
