% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fama_French_Download.R
\name{ff3factor}
\alias{ff3factor}
\title{Download Fama-French 3 Factor Model Data}
\usage{
ff3factor()
}
\value{
famafrench3factor
}
\description{
Function to download data from Ken French Website for 3 Factor Model.
}
\details{
Download
}
