% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fama_French_Download.R
\name{ff5factor}
\alias{ff5factor}
\title{Download Fama-French 5 Factor Model Data}
\usage{
ff5factor()
}
\value{
famafrench3factor
}
\description{
Function to download data from Ken French Website for 5 Factor Model.
}
\details{
Download
}
