% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fizz_import.R
\name{Fizz_import}
\alias{Fizz_import}
\title{Imports and Decode Fizz Format Excel Sheets}
\usage{
Fizz_import(file_path)
}
\arguments{
\item{file_path}{The path to the file to import}
}
\value{
The file restructured and decoded
}
\description{
This function imports and decodes Fizz files to be directly used in R or in Shiny Apps.
}
\examples{
NULL
}
