% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{formatting}
\alias{formatting}
\title{Wrapper for Formatting Numerical Values}
\usage{
formatting(x, n)
}
\arguments{
\item{x}{Numerical value or vector of numerical value to formatter}

\item{n}{Number of decimals after the comma.}
}
\value{
Vector of character of the same length as x, with values formatted with the right number of decimals
}
\description{
Wrapper to format numerical values to the right number of decimals. If the value is missing, they are replaced with an empty string.
}
\examples{
x = c(1.6589, NA, 0.15681, 32)
formatting(x, n=2)
}
