% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utilities.R
\name{get_args}
\alias{get_args}
\title{Get the argument string from an input-file line}
\usage{
get_args(infile, line_num)
}
\arguments{
\item{infile}{A tibble created by \code{\link[=get_input_file]{get_input_file()}}, or a character vector.}

\item{line_num}{Line number in the input file.}
}
\value{
Everything after the first colon; \code{NA_character_} if the line holds
no colon or does not exist.
}
\description{
The arguments are everything after the first colon; splitting them into
individual "parameter = value" pairs is \code{\link[=parse_args]{parse_args()}}'s job.
}
