% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utilities.R
\name{get_dem}
\alias{get_dem}
\title{Get the DEM raster files listed in an ASCII input file}
\usage{
get_dem(infile)
}
\arguments{
\item{infile}{A tibble created by \code{\link[=get_input_file]{get_input_file()}}.}
}
\value{
A list of DEM file names, one per DEM keyword in the file.
}
\description{
Get the DEM raster files listed in an ASCII input file
}
