% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utilities.R
\name{get_keyword}
\alias{get_keyword}
\title{Get the keyword from an input-file line}
\usage{
get_keyword(infile, line_num)
}
\arguments{
\item{infile}{A tibble created by \code{\link[=get_input_file]{get_input_file()}}, or a character vector.}

\item{line_num}{Line number in the input file.}
}
\value{
The keyword, trimmed of surrounding whitespace; \code{NA_character_} if
the line is a comment, is blank, or contains no colon.
}
\description{
The keyword is everything before the first colon on a line that is not a
comment.
}
