% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utilities.R
\name{get_raster_list}
\alias{get_raster_list}
\title{Parse a RASTER LIST block out of an input file}
\usage{
get_raster_list(
  infile,
  start_keyword = "RASTER LIST",
  end_pattern = "^END( RASTER)? LIST$"
)
}
\arguments{
\item{infile}{A tibble created by \code{\link[=get_input_file]{get_input_file()}}.}

\item{start_keyword}{Keyword opening the block.}

\item{end_pattern}{Regular expression matching the keyword that closes it.}
}
\value{
A list of \code{\link[=raster_list_entry]{raster_list_entry()}} objects, empty if the file has no
raster list.
}
\description{
Parse a RASTER LIST block out of an input file
}
