% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_utils.R
\name{log_module_access}
\alias{log_module_access}
\title{Log module access
Call this when a user accesses a specific module}
\usage{
log_module_access(session, module_name, log_path = "logs/usage_log.csv")
}
\arguments{
\item{session}{Shiny session object}

\item{module_name}{Name/ID of the module being accessed}

\item{log_path}{Path to the log file}
}
\description{
Log module access
Call this when a user accesses a specific module
}
