% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_utils.R
\name{log_session_end}
\alias{log_session_end}
\title{Log session end
Call this when a user disconnects from the app}
\usage{
log_session_end(session, log_path = "logs/usage_log.csv")
}
\arguments{
\item{session}{Shiny session object}

\item{log_path}{Path to the log file}
}
\description{
Log session end
Call this when a user disconnects from the app
}
