% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_utils.R
\name{log_session_start}
\alias{log_session_start}
\title{Log session start
Call this when a user connects to the app}
\usage{
log_session_start(session, log_path = "logs/usage_log.csv")
}
\arguments{
\item{session}{Shiny session object}

\item{log_path}{Path to the log file}
}
\description{
Log session start
Call this when a user connects to the app
}
