% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_metric_distribution}
\alias{plot_metric_distribution}
\title{Plot parameters by sample points}
\usage{
plot_metric_distribution(training_data, parameters = NULL, plot_type = "hist")
}
\arguments{
\item{training_data}{A data frame with the data to be plotted. Must have at
least a column called "class" of type \if{html}{\out{<chr>}} that contains a value "positive"
or "negative" for each row.}

\item{parameters}{A list of names of the parameters to plot. These must
correspond to columns in training_data}
}
\description{
Generates histogram plot comparing the distribution of various
parameters across positive and negative classes of data. This function takes
training_data as a data frame.
}
