% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_data_files.R
\name{recreate_data_files}
\alias{recreate_data_files}
\title{Recreate original data file(s) exactly as uploaded}
\usage{
recreate_data_files(file_inputs, output_dir = tempdir())
}
\arguments{
\item{file_inputs}{Single file input or list of file inputs from Shiny}

\item{output_dir}{Directory where to save the recreated files}
}
\value{
Character vector of created file paths
}
\description{
Recreate original data file(s) exactly as uploaded
}
\examples{
NULL
}
