% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_inputs_snapshot.R
\name{save_inputs_snapshot}
\alias{save_inputs_snapshot}
\title{Save input snapshot to Excel}
\usage{
save_inputs_snapshot(
  inputs_tibble,
  filepath,
  add_metadata = TRUE,
  app_name = "Shiny App"
)
}
\arguments{
\item{inputs_tibble}{Tibble from capture_shiny_inputs()}

\item{filepath}{Path where to save the Excel file}

\item{add_metadata}{Logical, whether to add a metadata sheet}

\item{app_name}{Name of the application (for metadata)}
}
\value{
Something
}
\description{
Save input snapshot to Excel
}
\examples{
NULL
}
