% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utilities.R
\name{strip_flt}
\alias{strip_flt}
\title{Remove a trailing ".flt" from a raster name}
\usage{
strip_flt(in_raster)
}
\arguments{
\item{in_raster}{Raster file name.}
}
\value{
The name without its ".flt" extension.
}
\description{
Remove a trailing ".flt" from a raster name
}
