% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableformat.R
\name{tableformat}
\alias{tableformat}
\title{Format Table for PowerPoint}
\usage{
tableformat(table)
}
\arguments{
\item{table}{The table to be re-formated}
}
\value{
The table as a formatted flextable
}
\description{
Transform R tables to pre-formated flextable before export to PowerPoint
}
\examples{
NULL
}
