% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfill_vec.R
\name{unfill_vec}
\alias{unfill_vec}
\title{Unfill Vector}
\usage{
unfill_vec(x)
}
\arguments{
\item{x}{The vector of elements to evaluated}
}
\value{
A vector of a same length as x, but with following identical values being replaced by ""
}
\description{
Opposite function to the fill_vec function: when following rows contain the same value, it only keeps the first one until a new value is observed, the repeated values being replaced by an empty cell ("")
}
\examples{
x = tibble::tibble(Col1 = c(rep("A",5),rep("B",3)), Col2 = c(1:5,1:3))
dplyr::mutate(x, Col1 = unfill_vec(Col1))
}
