% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_utilities.R
\name{write_attribute_list}
\alias{write_attribute_list}
\title{Write a whole ATTRIBUTE LIST block}
\usage{
write_attribute_list(
  writer,
  attribute_list,
  indent = 2L,
  end_keyword = "END LIST"
)
}
\arguments{
\item{writer}{A writer list from \code{\link[=input_writer]{input_writer()}}.}

\item{attribute_list}{List of \code{\link[=attribute_spec]{attribute_spec()}} objects.}

\item{indent}{Leading spaces for the ATTRIBUTE LIST and end keywords.}

\item{end_keyword}{Keyword closing the block. RIL uses "END LIST"; netrace
uses "END ATTRIBUTE LIST".}
}
\value{
The number of attributes written, invisibly.
}
\description{
Write a whole ATTRIBUTE LIST block
}
